# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Delibasket'
  version: ''
paths:
...
  /api/recipes/:
    get:
      operationId: Список рецептов
      description: Страница доступна всем пользователям. Доступна фильтрация по избранному, автору, списку покупок и тегам.
      parameters:
        - name: page
          required: false
          in: query
          description: Номер страницы.
          schema:
            type: integer
        - name: limit
          required: false
          in: query
          description: Количество объектов на странице.
          schema:
            type: integer
        - name: is_favorited
          required: false
          in: query
          description: Показывать только рецепты, находящиеся в списке избранного.
          schema:
            type: integer
            enum: [0, 1]
        - name: is_in_shopping_cart
          required: false
          in: query
          description: Показывать только рецепты, находящиеся в списке покупок.
          schema:
            type: integer
            enum: [0, 1]
        - name: author
          required: false
          in: query
          description: Показывать рецепты только автора с указанным id.
          schema:
            type: integer
        - name: tags
          required: false
          in: query
          description: Показывать рецепты только с указанными тегами (по slug)
          example: 'lunch&tags=breakfast'

          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                    description: 'Общее количество объектов в базе'
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://delibasket.example.org/api/recipes/?page=4
                    description: 'Ссылка на следующую страницу'
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://delibasket.example.org/api/recipes/?page=2
                    description: 'Ссылка на предыдущую страницу'
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipeList'
                    description: 'Список объектов текущей страницы'
          description: ''
      tags:
        - Рецепты
    post:
      security:
        - Token: []
      operationId: Создание рецепта
      description: 'Доступно только авторизованному пользователю'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreateUpdate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: 'Рецепт успешно создан'
        '400':
          description: 'Ошибки валидации в стандартном формате DRF'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/schemas/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
...
  /api/recipes/{id}/:
    get:
      operationId: Получение рецепта
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: ''
      tags:
        - Рецепты
    patch:
      operationId: Обновление рецепта
      security:
        - Token: [ ]
      description: 'Доступно только автору данного рецепта'
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта."
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreateUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: 'Рецепт успешно обновлен'
        '400':
          $ref: '#/components/responses/NestedValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
    delete:
      operationId: Удаление рецепта

      description: 'Доступно только автору данного рецепта'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '204':
          description: 'Рецепт успешно удален'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
  /api/recipes/{id}/favorite/:
    post:
      operationId: Добавить рецепт в избранное
      description: 'Доступно только авторизованному пользователю.'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeMinified'
          description: 'Рецепт успешно добавлен в избранное'
        '400':
          description: 'Ошибка добавления в избранное (Например, когда рецепт уже есть в избранном)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfMadeError'
        '401':
          $ref: '#/components/responses/AuthenticationError'

      tags:
        - Избранное
    delete:
      operationId: Удалить рецепт из избранного
      description: 'Доступно только авторизованным пользователям'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта."
          schema:
            type: string
      responses:
        '204':
          description: 'Рецепт успешно удален из избранного'
        '400':
          description: 'Ошибка удаления из избранного (Например, когда рецепта там не было)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfMadeError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Избранное
...
  /api/ingredients/:
    get:
      operationId: Список ингредиентов
      description: 'Список ингредиентов с возможностью поиска по имени.'
      parameters:
        - name: name
          required: false
          in: query
          description: Поиск по частичному вхождению в начале названия ингредиента.
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
          description: ''
      tags:
        - Ингредиенты
  /api/ingredients/{id}/:
    get:
      operationId: Получение ингредиента
      description: 'Уникальный идентификатор этого ингредиента.'
      parameters:
        - name: id
          in: path
          required: true
          description: ''
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
          description: ''
      tags:
        - Ингредиенты
...
  /api/auth/token/login/:
    post:
      operationId: Получить токен авторизации
      description: Используется для авторизации по емейлу и паролю, чтобы далее использовать токен при запросах.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenGetResponse'
          description: ''
      tags:
        - Пользователи
...
components:
  schemas:
    User:
      description:  'Пользователь (В рецепте - автор рецепта)'
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: "Адрес электронной почты"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Вася"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Пупкин"
        is_subscribed:
          type: boolean
          readOnly: true
          description: "Подписан ли текущий пользователь на этого"
          example: false
      required:
        - username
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          description: 'Название'
          example: 'Завтрак'
        color:
          type: string
          nullable: true
          maxLength: 7
          description: 'Цвет в HEX'
          example: '#E26C2D'
        slug:
          type: string
          nullable: true
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
          description: 'Уникальный слаг'
          example: 'breakfast'
    RecipeList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        tags:
          description: 'Список тегов'
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        author:
          $ref: '#/components/schemas/User'
        ingredients:
          description: 'Список ингредиентов'
          type: array
          items:
            $ref: '#/components/schemas/IngredientInRecipe'
        is_favorited:
          type: boolean
          description: 'Находится ли в избранном'
        is_in_shopping_cart:
          type: boolean
          description: 'Находится ли в корзине'
        name:
          type: string
          maxLength: 200
          description: 'Название'
        image:
          description: 'Ссылка на картинку на сайте'
          example: 'http://delibasket.example.org/media/recipes/images/image.jpeg'
          type: string
          format: url
        text:
          description: 'Описание'
          type: string
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
      required:
        - tags
        - author
        - is_favorited
        - is_in_shopping_cart
        - name
        - image
        - text
        - cooking_time
    RecipeMinified:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        name:
          type: string
          maxLength: 200
          description: 'Название'
        image:
          description: 'Ссылка на картинку на сайте'
          example: 'http://delibasket.example.org/media/recipes/images/image.jpeg'
          type: string
          format: url
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          example: 'Капуста'
        measurement_unit:
          type: string
          maxLength: 200
          example: 'кг'
      required:
        - name
        - measurement_unit
    IngredientInRecipe:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          description: 'Название'
          example: 'Картофель отварной'
        measurement_unit:
          type: string
          maxLength: 200
          description: 'Единицы измерения'
          example: 'г'
        amount:
          type: integer
          description: 'Количество'
          minimum: 1

      required:
        - name
        - measurement_unit
...
    TokenCreate:
      type: object
      properties:
        password:
          type: string
        email:
          type: string
    TokenGetResponse:
      type: object
      properties:
        auth_token:
          type: string
    RecipeCreateUpdate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        ingredients:
          description: Список ингредиентов
          type: array
          items:
            example:
              id: 1123
              amount: 10
            type: object
            properties:
              id:
                description: 'Уникальный id'
                type: integer
              amount:
                description: 'Количество в рецепте'
                type: integer
            required:
              - id
              - amount
        tags:
          description: 'Список id тегов'
          type: array
          example: [1, 2]
          items:
            type: integer
        image:
          description: 'Картинка, закодированная в Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
        name:
          description: 'Название'
          type: string
          maxLength: 200
        text:
          description: 'Описание'
          type: string
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
      required:
        - ingredients
        - tags
        - image
        - name
        - text
        - cooking_time
    ValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        field_name:
          description: 'Название поля, в котором произошли ошибки. Таких полей может быть несколько'
          example: [ 'Обязательное поле.' ]
          type: array
          items:
            type: string
    NestedValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        ingredients:
          description: 'Ошибки в ингредиентах. В приведенном примере в первом и третьем ингредиенте не было ошибок (amount >= 1), а во втором были.'
          example: [{},{"amount":["Убедитесь, что это значение больше либо равно 1."]}, {}]
          type: array
          items:
            type: object
            properties:
              amount:
                type: array
                items:
                  type: string
    SelfMadeError:
      description: Ошибка
      type: object
      properties:
        errors:
          description: 'Описание ошибки'
          type: string
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
    PermissionDenied:
      description: Недостаточно прав
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "У вас недостаточно прав для выполнения данного действия."
          type: string
    NotFound:
      description: Объект не найден
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
  responses:
...
    NestedValidationError:
      description: 'Ошибки валидации в стандартном формате DRF, в том числе с внутренними элементами.'
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/NestedValidationError'
              - $ref: '#/components/schemas/ValidationError'
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token